+++ Types/Turtle/Towards.ps1
@@ -2,7 +2,7 @@
 .SYNOPSIS
     Determines the angle towards a point
 .DESCRIPTION
-    Determines the angle from the turtle's current position towards a point.
+    Determines the angle from the turtle's current heading towards a point.
 #>
 param(
 # The X-coordinate
@@ -12,9 +12,9 @@ param(
 )
 
 # Determine the delta from the turtle's current position to the specified point
-$deltaX = $this.Position.X - $X
-$deltaY = $this.Position.Y - $Y
+$deltaX = $X - $this.Position.X 
+$deltaY = $Y - $this.Position.Y
 # Calculate the angle in radians and convert to degrees
 $angle = [Math]::Atan2($deltaY, $deltaX) * 180 / [Math]::PI
-# Return the angle rotate by 90 to account for the turtle's coordinate system
-return $angle + 90
+# Return the angle minus the current heading (modulo 360)
+return $angle - ($this.Heading % 360)