+++ CHANGELOG.md
@@ -1,3 +1,24 @@
+## Turtle 0.1.4
+
+* `Turtle` Upgrades
+  * `turtle` will return an empty turtle (#112)
+  * `turtle` now splats to script methods, enabling more complex input binding (#121)
+  * `LSystem` is faster and more flexible (#116)
+* New Properties:
+  * `get/set_Opacity` (#115)
+  * `get/set_PathAnimation` (#117)
+  * `get/set_Width/Height` (#125)
+* New Methods:
+  * `HorizontalLine/VerticalLine` (#126)
+  * `Petal` (#119)
+  * `FlowerPetal` (#124)
+  * `Spirolateral` (#120)
+  * `StepSpiral` (#122)
+* Fixes:
+  * `Turtle.Towards()` returns a relative angle (#123)
+
+---
+
 ## Turtle 0.1.3
 
 * Fixing `Get-Turtle` inline sets (#108, #107)